Added last updates into `ConfigurationManager`.
