+++ lib/base-data-service/base-api-service.js
@@ -1,5 +1,5 @@
 import { ApiAdapter } from 'lib/api-adapter';
-import { ConfigurationManager } from 'lib/configuration-manager/ConfigurationManager';
+import { Configuration } from 'lib/configuration-manager/Configuration';
 
 
 
@@ -23,7 +23,7 @@ export class BaseApiService {
       throw new TypeError('[BaseApiService] Unrecognized `apiAdapter` argument, it must be the instance of `ApiAdapter`');
     }
 
-    if (! (config instanceof ConfigurationManager)) {
+    if (! (config instanceof Configuration)) {
       throw new TypeError('[BaseApiService] Unrecognized `config` argument, it must be the instance of `ConfigurationManager`');
     }
