+++ lib/base-data-service/base-data-service.js
@@ -1,6 +1,6 @@
 import { iocFactory } from 'libioc';
 import { BaseApiService } from './base-api-service';
-import { ConfigurationManager } from 'lib/configuration-manager/ConfigurationManager';
+import { Configuration } from 'lib/configuration-manager/Configuration';
 
 
 
@@ -44,7 +44,7 @@ export class BaseDataService {
       throw new TypeError('[BaseDataService] Unrecognized `apiService` argument, it must be the instance of `BaseApiService`');
     }
 
-    if (! (config instanceof ConfigurationManager)) {
+    if (! (config instanceof Configuration)) {
       throw new TypeError('[BaseDataService] Unrecognized `config` argument, it must be the instance of `ConfigurationManager`');
     }
