+++ lib/configuration-manager-v1/ServerConfigurationManager.js
