+++ lib/configuration-manager-v1/readers/BaseConfigurationReader.js
