+++ lib/configuration-manager-v1/readers/EnvFileConfigurationReader.js
