+++ lib/configuration-manager/Configuration/Configuration.js
