+++ lib/configuration-manager/Configuration/Configuration.test.js
