+++ lib/configuration-manager/Configuration/index.js
@@ -1,0 +1,1 @@
+export { Configuration } from './Configuration';
