+++ lib/configuration-manager/ConfigurationManager.js
@@ -1,25 +1,54 @@
+import path from 'path';
+import parseArgs from 'minimist';
+import { createReaderByType } from './readers';
+import { ServerConfiguration } from './Configuration/ServerConfiguration';
+
+
 /**
- * Encapsulate work with configuration data
- * Provides interface for getting nested config data via "dot notation" (e.g. "service.api.endpoint")
+ * Responsive to handle configuration specific operation, organized between config reader and access object
  */
 export class ConfigurationManager {
-  constructor(configurationData) {
-    this.config = configurationData;
 
-    // freeze config data, since config data must be immutable during app lifecycle
-    Object.freeze(this.config);
+  static parseEnvVars() {
+    return Object.assign({}, process.env);
   }
 
-  get(dotNotation) {
-    const dotTokens = dotNotation.split('.');
 
-    try {
-      return dotTokens.reduce(
-        (res, token) => res[token],
-        this.config
-      );
-    } catch (e) {
-      throw new Error(`ConfigurationManager: key ${dotNotation} was not found in configuration`);
-    }
+
+  static parseCliArgs() {
+    return parseArgs(process.argv.slice(2));
+  }
+
+
+  constructor(rootDir) {
+    this.rootDir = rootDir;
+  }
+
+
+
+  initialize({ readerType, appName }) {
+    return new Promise((resolve, reject) => {
+      const envVars = ConfigurationManager.parseEnvVars();
+      const cliArgs = ConfigurationManager.parseCliArgs();
+
+      // load default config
+      const fileReader = createReaderByType('file');
+      const config = new ServerConfiguration(fileReader.readDefaultConfigFileSync(path.join(this.rootDir, 'config')));
+
+      // load specific config
+      const reader = createReaderByType(readerType);
+
+      return reader.read({ appName, envVars, cliArgs })
+        .then(configData => {
+          // merge default config with specific
+          config.mergeWith(configData);
+
+          // load "local" config file
+          config.mergeWith(fileReader.readLocalConfigFileSync(path.join(this.rootDir, 'config')));
+
+          resolve(config);
+        })
+        .catch(err => reject(err));
+    });
   }
 }
