+++ lib/configuration-manager/index.js
@@ -1,0 +1,2 @@
+export { ConfigurationManager } from './ConfigurationManager';
+export { Configuration } from './Configuration';
