+++ lib/configuration-manager/readers/FileConfigurationReader.js
@@ -69,6 +69,10 @@ export class FileConfigurationReader extends IConfigurationReader {
     return new Promise((resolve, reject) => {
       const configDirPath = cliArgs['config-dir'];
 
+      if (! configDirPath) {
+        reject(new Error('[FileConfigurationReader] `--config-dir` cli argument isn\'t provided'));
+      }
+
       try {
         const configData = FileConfigurationReader.readConfigFile(path.join(configDirPath, appName));
