+++ lib/configuration-manager/readers/IConfigurationReader.js
