+++ lib/configuration-manager/readers/index.js
@@ -1,24 +1,17 @@
-import { EnvFileConfigurationReader } from './EnvFileConfigurationReader';
+import { FileConfigurationReader } from './FileConfigurationReader';
+
+const readersCollection = [
+  FileConfigurationReader
+];
 
 
 
-/**
- * Responsible to instantiate configuration reader according to supplied `type` argument
- * @param {String} type
- * @param {String} configReaderParams
- * @returns {BaseConfigurationReader}
- */
-export function configReaderFactory(type, configReaderParams) {
+export function createReaderByType(type) {
+  const TargetReader = readersCollection.find(({ readerType }) => readerType === type);
 
-  switch (type) {
-    case EnvFileConfigurationReader.readerType:
-      return new EnvFileConfigurationReader(configReaderParams);
-
-    default:
-      throw new TypeError(`configReaderFactory: Unrecognized reader type "${type}"`);
+  if (! TargetReader) {
+    throw new TypeError(`ConfigurationManager: Unrecognized reader type "${type}"`);
   }
+
+  return new TargetReader;
 }
-
-
-configReaderFactory.$singleton = true;
-configReaderFactory.$inject = ['CONFIG_READER_TYPE', 'CONFIG_READER_PARAMS'];
