+++ lib/ui-application/client.ioc-container.js
@@ -1,5 +1,5 @@
 import { IoCContainer, iocFactory, iocClass } from 'libioc';
-import { ConfigurationManager } from 'lib/configuration-manager/ConfigurationManager';
+import { Configuration } from 'lib/configuration-manager/Configuration';
 import { DummyLogger } from '../dummy-logger';
 import { createApiAdapter } from '../api-adapter';
 import { createAxiosTransport } from '../api-adapter/transports/AxiosTransport';
@@ -25,7 +25,7 @@ export function createClientIoCContainer(configData) {
     logger: iocClass(DummyLogger).asSingleton(),
 
     // App configuration
-    config: iocFactory(() => new ConfigurationManager(configData)).asSingleton(),
+    config: iocFactory(() => new Configuration(configData)).asSingleton(),
 
     // Api Adapter
     httpTransport: iocFactory(createAxiosTransport),
