+++ lib/ui-application/server.ioc-container.js
@@ -1,6 +1,4 @@
 import { IoCContainer, iocClass, iocFactory } from 'libioc';
-import { configReaderFactory } from 'lib/configuration-manager/readers';
-import { serverConfigurationFactory } from 'lib/configuration-manager/ServerConfigurationManager';
 import { DummyLogger } from '../dummy-logger';
 import { createApiAdapter } from '../api-adapter';
 import { createNodeRequestTransport } from '../api-adapter/transports/NodeRequestTransport';
@@ -13,23 +11,19 @@ import { AssetsManifestManager } from '../assets-manifest-manager';
 /**
  * Initial server IoC container with all bootstrap dependencies
  *
- * @param {String} configReaderType - type of concrete config reader
- * @param {Object} configReaderParams - params intended to be passed to concrete config reader constructor
+ * @param {ServerConfiguration} config
  * @param {Object} assetsManifestPath
  * @returns {IoCContainer}
  */
-export function createServerIocContainer({ configReaderType, configReaderParams, assetsManifestPath }) {
+export function createServerIocContainer({ config, assetsManifestPath }) {
   const iocContainer = new IoCContainer;
 
   iocContainer.registerAll({
     // some deps rely on IoC itself, so due to that we need to register IoC instance as well
     ioc: iocContainer,
 
-    // register config specific entities
-    CONFIG_READER_TYPE: configReaderType,
-    CONFIG_READER_PARAMS: configReaderParams,
-    configReader: iocFactory(configReaderFactory),
-    config: iocFactory(serverConfigurationFactory),
+    // register config
+    config,
 
     // manifest specific entities
     MANIFEST_PATH: assetsManifestPath,
